requirements.txt
keyboard>=0.13.5
requests==2.31.0
beautifulsoup4>=4.12.2
pandas>=2.1.1
selenium>=4.14.0
python-dotenv>=1.0.0
undetected-chromedriver>=3.5.0
selenium-stealth>=1.0.6
fake-useragent>=1.1.1
cloudscraper>=1.2.71
playwright==1.36.0
playwright-stealth>=1.0.5
puppeteer-stealth>=1.0.0
httpx>=0.24.1
websockets==11.0.3
cryptography==41.0.1
requests-toolbelt==1.0.0
tenacity==8.2.2
jwt==4.0.0
aiohttp>=3.8.5
opencv-python>=4.8.0
pytesseract>=0.3.10
numpy==1.24.3
pillow>=10.0.1
mss>=9.0.1
python-opencv>=4.8.0
easyocr>=1.7.1
schedule>=1.2.0
lxml>=4.9.2
webdriver-manager>=3.8.6
plotly>=5.14.1
psutil>=5.9.5
gputil>=1.4.0
scikit-learn>=1.3.2
tensorflow>=2.14.0
torch>=2.1.0
xgboost>=2.0.0
lightgbm>=4.1.0
optuna>=3.4.0
statistics==1.0.3.5
ephem==4.1.4

# System Agent Dependencies
psutil>=5.9.0
pyautogui>=0.9.53
selenium>=4.14.0
beautifulsoup4>=4.12.2
transformers>=4.36.0
torch>=2.1.0

# Database and Storage
sqlalchemy>=2.0.0
sqlite3-api>=0.1.0
pymongo>=4.6.0
redis>=5.0.0

# Astronomical Calculations
playwright==1.36.0
ephem==4.1.4
numpy==1.24.3
pandas==2.1.4
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3.post1
tenacity==8.2.2
cryptography==41.0.1
requests-toolbelt==1.0.0
